Create Bobby
You clearly didn't understand the algorithm or how to set up a function.
Your algorithm will never work as a setup. Look at your logic. if 'a' = 0 then c = (a * b) will always be 0
and therefore so will b since b = c.

As for your function. You were to pass it an argument and output a value. Yours does neither.

Lastly, the whole thing was supposed to be in a loop so that the user can enter repeated requests.